# soya-framework-admin/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "Workshop Server",
    "contact" : {
      "name" : "wenqun.soya@gmail.com"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/cmd/command/{group}/{command}" : {
      "get" : {
        "tags" : [ "Command Service" ],
        "operationId" : "command",
        "consumes" : [ "text/plain" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "group",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "command",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/resource/aes-decrypt" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "aesDecrypt",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "secret",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/unzip" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "unzip",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/aes-encrypt" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "aesEncrypt",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "secret",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/base64-decode" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "base64Decode",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/base64-encode" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "base64Encode",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/json-format" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "jsonFormat",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/echo" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "echo",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/extract" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "extract",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/mustache-attribute" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "mustacheAttribute",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/resource/gzip" : {
      "post" : {
        "tags" : [ "resource" ],
        "operationId" : "gzip",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  }
}